requirements.txt
# Discord bot framework
discord.py>=2.3.0

# Google Sheets API client
gspread>=6.0.0

# Google authentication
google-auth>=2.28.0

# .env file loading
python-dotenv>=1.0.0
